Tagging: Add get_tags_dir()
get_tags_dir() takes the COnstant TAGS_DIR and creates it before
returning it. If it exists, it does nothign and returns the TAGS_DIR.
If it is unable to create the TAGS_DIR, it returns None. It also gives
an error message stating that coala will continue without tagging.

Fixes https://github.com/coala-analyzer/coala/issues/1050